NLP_Feature_Extraction_and_Product_Categorization/requirements.txt
# Core
pandas>=2.1.0
numpy>=1.26.0
lxml>=5.1.0
pyyaml>=6.0.0

# NLP
spacy>=3.7.0

# ML & Feature Engineering
scikit-learn>=1.4.0
scipy>=1.12.0
xgboost>=2.0.0

# Visualization
plotly>=5.18.0
kaleido>=0.2.1
wordcloud>=1.9.0
matplotlib>=3.8.0

# Notebook
nbformat>=4.2.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0

# Data I/O
pyarrow>=15.0.0
